Added completed files via upload
Simulates a multithreaded, multi-process system with mutual exclusion through the use of a semaphore. For this example, the BridgeExclusion copy.java file is the semaphore , with the car copy.java file representing the threaded processes. SynchronizedTraffic.java and SynchronizedTrafficTester copy.java are used to test and implement the code. (files are a copy as the original files were placed elsewhere on the disk for terminal compilation and to be submitted).